requirements.txt
accelerate==1.1.1
fastapi==0.115.5
google-auth==2.36.0
google-generativeai==0.8.3
huggingface-hub==0.26.2
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-google-genai==2.0.4
langchain-huggingface==0.1.2
langchain-openai==0.2.8
pyowm
python-dotenv==1.0.1
sentencepiece==0.2.0
streamlit
torch==2.5.1 --index-url https://download.pytorch.org/whl/cu118
torchaudio==2.5.1
torchvision==0.20.1
transformers==4.46.2
uvicorn==0.32.0
wikipedia